Use ServiceBindingProjection to manage projection
We previously had two reconcilers managing the ServiceBinding resource:
the first to lookup the service secret and create the projected secret,
and the second to inject the projection into the target application.
Now, those two reconcilers reconciler the ServiceBinding and
ServiceBindingProjection resources respectively. the
ServiceBindingProjection resource is created by the ServiceBinding
reconciler.

While the ServiceBindingProjection resource is intended for use with
custom projections, we do not the support them (yet).

Note: the ServiceBindingProjection resource is an internal resource and
should not be created directly.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>