Support multiple bindings to the same application (#94)
Multiple bindings previously mostly worked, but it was inelegant.
Knative bindings uses a combination of a webhook and a controller to
manage the binding on the subject. The controller previously supported
multiple bindings while the webhook only applied one of the bindings.

This discrepancy created extra churn on the API Server as the controller
needed to mutate the resource immediately after it was created/updated.
In resource that stamp out new children, the intermediate state would
remain visible. Worse, immutable resources can't be updated, leaving a
half-baked resource stuck in perpetuity.

This issue has been fixed upstream, now the webhook applies all bindings
targeting an application. There is a new sample demonstrating multiple
bindings on an immutable resource.

Bump minimum k8s version to 1.17

Signed-off-by: Scott Andrews <andrewssc@vmware.com>