Drop Custom Projection support
The Custom Projection extension was dropped from the Service Binding
spec. We no longer need to maintain it.

The ServiceBindingProjection resource is still used to manage the
binding, but it has moved into the internal.bindings.labs.vmware.com API
group and is now an implementation detail rather than a supported
interface.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>