Test for clean redeploys with kapp
A kapp deploy will attempt to make the resources on the api server match
the resources defined locally. Because the manager will modify the
resources we create at runtime, we should respect those mutations and
avoid overwriting them. This is typically done with rebase rules in the
kapp config resource.

During the acceptance tests, we set the kapp-controller sync period for
the package to 1 minute. If we see multiple changeset, the config is not
stable.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>