Fix Webhook failures in k8s 1.22+
Kubernetes 1.22 added a new subresource field https://github.com/kubernetes/kubernetes/pull/100970. The knative pkg which is used the webhook had an issue that was resolved here https://github.com/knative/serving/issues/11448.

This change bumps the knative pkg dependency which includes the fix but keeps it pinned on the 0.22 release branch.

In addition it adds 1.22.6 to the k8s testing matrix in Github actions.

fixes #214